Header animation
These changes add the jQuery UI library, which allows us to use additional animations for the landing page or any other part of the site. Calls to old and new animations were chained with .stop() to allow for any current callback to stop if another takes precedence. For instance, if we are in an animation to go to the bottom of the page and we call another to go to the middle, we will stop the current callback to the bottom and go straight to the middle.