Patched header selection with scrolling events
The current header is designed to show what section of the landing we are on. This would only work if the links were clicked, so it would show that we are at the top of the page while we scroll to another part. I checked the current location of the scrollbar and compared it to the location of all sections in order to fix this.